Removed: UserLoadAfterLoadFromSession and UserLogoutComplete hooks
This forces all authentication to be done through Special:Connect. Also, clicking Logout doesn't log the user out of Facebook, just MediaWiki. The reason for this is that logging out of Facebook requires an extra webpage to be visited, and is unpractical as the user would then have to log in to Facebook.com again to use Facebook