[Changed] Move FacebookUserException back to FacebookUser.php.
This is a follow-up commit to 591e6ec4a.

Originally, FacebookUserException was renamed to FacebookModelException and moved to SpecialConnect.php to provide a generic exception for FacebookUser and FacebookApplication. However, FacebookApplication is a read-only object (writing happens via JavaScript). Thus, the exception isn't needed for the app class and can be renamed to FacebookUserException.