[Facebook session] Changed: Don't ping Facebook's servers every page load
This was originally implemented because I noticed a problem with stale sessions unable to fetch info, but still reporting the user as "logged in". The Facebook PHP SDK recently fixed a bug related to this (see b0b43b8) so I'm assuming this problem has gone away. If not, I'll revert this commit