[Open Graph] Added: "action" attribute for <opengraph> tag.
Use it like this: <opengraph action="want">I want this!</opengraph>

The action must first be defined using $wgFbOpenGraphCustomActions, and the wiki page must be a compatible custom object.

This will render a dummy link to /PAGENAME#want. In the future, JavaScript will push the action to the user's Timeline