[Fixed] Erratic behavior on first page load
This behavior would redirect the user to Special:Connect, even though they were already connected. On Special:Connect, because the user is already logged in, they wouldn't see any useful information and there would be no "return to previous page" button. The only way to exit Special:Connect would be to explicitly click the back button in their browser.

Steps to reproduce the bug can be found in the comment in FacebookHooks.php.