[Open Graph] Removed support for custom objects and actions.
These features are only removed from the settings file. The code for both features remains.

This change is for the version bump to 4.0-final. Because Open Graph actions are not fully implemented (some JavaScript code still needs to be written), they will not make the cut for 4.0-final. Support for these features is being pushed back to a later release.

Development for custom objects actions will take place in the "opengraph" branch on GitHub and will be merged with master "when it's done".